requirements.txt
# API and server
fastapi
uvicorn
python-multipart

# Audio processing
pydub

# OpenAI integration
openai

# Environment variables
python-dotenv